indexer: check whether ref resolved to commit
If a reference can't be resolved to a commit, libgit2's
git_revparse_single() will return a non-zero status. This patch adds a
check for this status and prints to stderr if this happens.

Previously codesearch would just crash with a null object assertion, if,
for example, run against an empty repository with no branches or
commits.