Add explanatory comments to object & struct fields
Some of these are reasonably obvious, but some are a bit subtle. I
thought it would be helpful to record the understanding I've extracted
from the code (and be corrected if my understanding is wrong!).

It seemed particularly useful to note the effective lifetimes of the
things which are only used during index construction. (It might further
be useful to factor those out into separate builder classes of some
sort.)