Skip chunks that don't contain desired repos
When indexing, codesearch builds chunks in the order that repositories
are listed in the configuration. When searching, it examines chunks in
the same order. This means that it finds results from repos in roughly
the order repos listed in the configuration (modulo multiple threads
concurrently processing chunks).

Since repo_pat constraints are applied after the main suffix array
search, this means that the search may examine and discard many
candidate lines before it ever reaches the desired repo, which can take
a long time. On our instance (17 repos, ~11GB index), restricting a
search like "e" to the last repo takes 6-7x as long as the same search
restricted to the first repo.

This change allows us to skip irrelevant chunks entirely. However,
the improvement in overall search performance is limited by the
unchunked filename search happening in parallel: even though the main
search completes more quickly, the RPC must still wait for the
filename search before returning results, so there's potential for
further improvement here.