Update “help” texts with recent file view features
This updates both the “help” popup as well as the help text across the
top of the file view header so that both match the complete list of new
text searching and selection features.  It also improves the markup and
CSS used to display the help popup with better tag semantics (“kbd”) and
visual display (lines are now uniformly spaced instead of having odd gaps).