Make template relationships explicit in their text
Instead of composing templates by first rendering the page template to a
string and then inserting it into the layout template, let’s load real
Go template stacks so that the relationship is explicit inside each page
template, and so that only a single rendering run is necessary.

By letting all templates render in the same data context, this will open
the way to CSP noncing of `<script>` tags whether they’re in the layout
or whether they live down in a page template.