Fix nonces by adding them to all script, link tags
It turns out?  Our CSP wants nonces on *every* script and link tag, not
just inline scripts.  Thanks to the recent template simplification, it’s
now easy to use the Nonce both up in the main “layout.html” template and
also down in specific templates like “fileview.html”.

This also simplifies the code by precomputing the full text of the HTML
attribute, so that the templates needn’t be festooned with “if” statements.