Update the name of the RE2 dependency
GRPC also depends on RE2, by the `com_github_google_re2` name. The
fact that we were using a different name resulted in GRPC pulling in a
second copy. This used to mostly work fine because they were
ABI-compatible, at least in the pieces we use. However, RE2 recently
updated the layout of a core struct in
572d6abf7f227053a4f8b83381fc4378714a2552, which meant that we ended up
seeing one version of the headers and linking against a different
version of the objects, which was a no good bad time.

IMO this is arguably a bug in GRPC's bazel build -- it should not be
leaking those headers -- but this fix is correct in any case.